server/requirements.txt
fastapi
uvicorn[standard]
cryptography
pydantic
requests
python-multipart
python-dateutil
